JAMES-1959 Extract JWT to an other maven project
