+++ server/pom.xml
@@ -109,6 +109,7 @@
         <module>protocols/webadmin-integration-test</module>
 
         <module>testing</module>
+        <module>protocols/jwt</module>
     </modules>
 
     <distributionManagement>
@@ -787,6 +788,11 @@
             <dependency>
                 <groupId>org.apache.james</groupId>
                 <artifactId>james-server-data-ldap-integration-testing</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.james</groupId>
+                <artifactId>james-server-jwt</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
