+++ server/protocols/jwt/src/main/java/org/apache/james/jwt/JwtTokenVerifier.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.jmap.crypto;
+package org.apache.james.jwt;
 
 import javax.inject.Inject;
