+++ server/protocols/jwt/src/test/java/org/apache/james/jwt/JwtTokenVerifierTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.jmap.crypto;
+package org.apache.james.jwt;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -24,7 +24,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.security.Security;
 import java.util.Optional;
 
-import org.apache.james.jmap.JMAPConfiguration;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.junit.Before;
 import org.junit.BeforeClass;
@@ -64,13 +63,9 @@ public class JwtTokenVerifierTest {
         sut = new JwtTokenVerifier(pubKeyProvider);
     }
 
-    private JMAPConfiguration getJWTConfiguration() {
+    private JwtConfiguration getJWTConfiguration() {
 
-        return JMAPConfiguration.builder()
-                .keystore(".")
-                .secret(".")
-                .jwtPublicKeyPem(Optional.ofNullable(PUBLIC_PEM_KEY))
-                .build();
+        return new JwtConfiguration(Optional.of(PUBLIC_PEM_KEY));
     }
 
     @Test
