+++ server/protocols/jwt/src/test/java/org/apache/james/jwt/PublicKeyReaderTest.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.jmap.crypto;
+package org.apache.james.jwt;
 
 import static org.assertj.core.api.Assertions.assertThat;
